Se cambian nombre de variables, se modifican pantallas, se separa bloques de codigo en metodos

+++ app/src/main/java/com/example/perraco/Objets/EstadisticaArticuloObject.kt
@@ -5,5 +5,5 @@ import java.io.Serializable
 data class EstadisticaArticuloObject(
     var _id:InventarioObjeto,
     var cantidad:Int,
-    var total:Long
+    var total:Double
 ): Serializable
